Added rope swing bp
In the all folder you can find BP_Rope. You can place it down and move it around where you like, to swing on it just jump into the rope and you'll be attached. Use W and S to swing back and forth. Press space to let go of the rope.