Set Up Boss AI
I started working on a basis for the Boss AI, included is: A blackboard, an AI controller, an interface for attacks, a behavior tree, 3 essential tasks, and a test of these blueprints which you can find in the ThirdPersonMap. Said test is just a placeholder until we establish what the bosses are going to be and what they're going to do.